feat(modules): add ShowModule mixin to various modules for improved module visibility

+++ filesystems/source/uim/filesystems/classes/metadata.d
@@ -66,7 +66,7 @@ FileMetadata getMetadata(string path) @trusted {
 }
 
 /// Get file size
-ulong getFileSize(string path) @trusted {
+size_t getFileSize(string path) @trusted {
     enforce(exists(path), "Path does not exist: " ~ path);
     return getSize(path);
 }
